medicare_project/requirements.txt
Django~=5.1.11
djangorestframework~=3.15
djangorestframework-simplejwt~=5.5
django-filter~=24.2
drf-spectacular~=0.27
django-cors-headers~=4.1
psycopg2-binary~=2.9
celery[redis]~=5.4
pytest-django~=4.8
